+++ CHANGELOG.md
@@ -12,9 +12,10 @@ Changes:
 - Add `.multi` support on `api.at(...).<section>.<method>`
 - Add support for ink! metadata V3 with payable constructors
 - Cleanup ink! metadata parsing, allowing for easier extension
+- Align `Range/RangeInclusive` handling with other definitions
 - Expose `contract.{query, tx}.<method>.meta`, aligning with API
 - Fix storage metadata, aligning method with decorated name
-- Adjust typegen, only using exportInterface
+- Adjust typegen, only use exportInterface
 - Added Kusama/Polkadot 9151 upgrade block (known types)
 - Update to latest Substrate, Kusama & Polkadot static metadata